feat: [AD-008] date time labels show on the bottom of component
