+++ view/src/components/AnomalyDots.tsx
@@ -83,62 +83,6 @@ export default function AnomalyDots({
     return gridLines;
   };
 
-  const generateDateLabels = () => {
-    const labels: React.ReactElement[] = [];
-
-    const startHour = new Date(minTime);
-    startHour.setMinutes(0, 0, 0);
-    const endHour = new Date(maxTime);
-    endHour.setMinutes(59, 59, 999);
-
-    const currentHour = new Date(startHour);
-    let index = 0;
-
-    while (currentHour <= endHour) {
-      const hourTime = currentHour.getTime();
-      if (hourTime >= minTime && hourTime <= maxTime) {
-        const leftPercent = getLeftPercent(hourTime);
-        const date = new Date(hourTime);
-
-        const dateLabel = date.toLocaleDateString([], {
-          month: "short",
-          day: "numeric",
-        });
-        const timeLabel = date.toLocaleTimeString([], {
-          hour: "2-digit",
-          minute: "2-digit",
-        });
-
-        labels.push(
-          <div
-            key={`date-${index}`}
-            style={{
-              position: "absolute",
-              left: `${leftPercent}%`,
-              bottom: -35,
-              transform: "translateX(-50%)",
-              fontSize: 9,
-              color: "#666",
-              whiteSpace: "nowrap",
-              pointerEvents: "none",
-              zIndex: 2,
-              textAlign: "center",
-              lineHeight: "1.2",
-            }}
-          >
-            <div>{dateLabel}</div>
-            <div>{timeLabel}</div>
-          </div>
-        );
-      }
-
-      currentHour.setHours(currentHour.getHours() + 1);
-      index++;
-    }
-
-    return labels;
-  };
-
   const generateAnomalyLabels = () => {
     const labels: React.ReactElement[] = [];
 
@@ -224,11 +168,9 @@ export default function AnomalyDots({
   return (
     <div
       ref={containerRef}
-      style={{ width: "100%", position: "relative", height, paddingBottom: 40 }}
+      style={{ width: "100%", position: "relative", height, paddingBottom: 10 }}
     >
       {generateTimeGridLines()}
-
-      {generateDateLabels()}
 
       {generateAnomalyLabels()}
